Rename `i18n-new` folder to `i18n`
