+++ server/i18n/makeCollectionTranslatable.ts
