+++ server/i18n/resourceSlugForCollection.js
