+++ server/i18n/returnNullIfFunctionThrows404.ts
